bug: 修复DDC finalize ref时仓库类型转换错误 #1161
