+++ src/backend/ddc/biz-ddc/src/main/kotlin/com/tencent/bkrepo/ddc/service/ReferenceService.kt
@@ -29,6 +29,8 @@ package com.tencent.bkrepo.ddc.service
 
 import com.tencent.bkrepo.common.api.exception.BadRequestException
 import com.tencent.bkrepo.common.api.message.CommonMessageCode
+import com.tencent.bkrepo.common.artifact.manager.StorageManager
+import com.tencent.bkrepo.common.artifact.repository.context.ArtifactContextHolder
 import com.tencent.bkrepo.common.security.util.SecurityUtils
 import com.tencent.bkrepo.ddc.config.DdcProperties
 import com.tencent.bkrepo.ddc.exception.ReferenceIsMissingBlobsException
@@ -40,6 +42,7 @@ import com.tencent.bkrepo.ddc.pojo.Reference
 import com.tencent.bkrepo.ddc.repository.RefRepository
 import com.tencent.bkrepo.ddc.serialization.CbObject
 import com.tencent.bkrepo.ddc.utils.hasAttachments
+import com.tencent.bkrepo.repository.api.NodeClient
 import org.bson.types.Binary
 import org.slf4j.LoggerFactory
 import org.springframework.stereotype.Service
@@ -52,6 +55,8 @@ class ReferenceService(
     private val blobService: BlobService,
     private val refResolver: ReferenceResolver,
     private val refRepository: RefRepository,
+    private val nodeClient: NodeClient,
+    private val storageManager: StorageManager,
 ) {
     fun create(ref: Reference): Reference {
         val inlineBlob = if (ref.inlineBlob!!.size > ddcProperties.inlineBlobMaxSize.toBytes()) {
@@ -83,19 +88,33 @@ class ReferenceService(
     }
 
     fun getReference(
-        refId: RefId,
-        includePayload: Boolean = true,
+        projectId: String,
+        repoName: String,
+        bucket: String,
+        key: String,
         checkFinalized: Boolean = true,
+        includePayload: Boolean = true,
     ): Reference? {
-        with(refId) {
-            val tRef = refRepository.find(projectId, repoName, bucket, key, includePayload) ?: return null
-            if (checkFinalized && !tRef.finalized) {
-                throw BadRequestException(
-                    CommonMessageCode.PARAMETER_INVALID, "Object ${tRef.bucket} ${tRef.key} is not finalized."
-                )
-            }
+        val tRef = refRepository.find(projectId, repoName, bucket, key, includePayload) ?: return null
+        if (checkFinalized && !tRef.finalized) {
+            throw BadRequestException(
+                CommonMessageCode.PARAMETER_INVALID, "Object ${tRef.bucket} ${tRef.key} is not finalized."
+            )
+        }
 
-            return Reference.from(tRef)
+        val ref = Reference.from(tRef)
+        if (ref.inlineBlob == null) {
+            val repo = ArtifactContextHolder.getRepoDetail(ArtifactContextHolder.RepositoryId(projectId, repoName))
+            ref.inlineBlob = nodeClient.getNodeDetail(projectId, repoName, ref.fullPath()).data?.let {
+                storageManager.loadArtifactInputStream(it, repo.storageCredentials)?.readBytes()
+            }
+        }
+
+        return if (ref.inlineBlob == null) {
+            logger.warn("Blob was null when attempting to fetch ${ref.repoName} ${ref.bucket} ${ref.key}")
+            null
+        } else {
+            ref
         }
     }
